hub-api/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
langgraph>=0.1.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
redis>=5.0.0
pinecone-client>=3.0.0
pgvector>=0.2.5
presidio-analyzer>=2.2.354
presidio-anonymizer>=2.2.354
spacy>=3.7.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
sse-starlette>=2.1.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
